Dodanie BasePresenter.kt i BaseContract.kt Przebudowa presenterów i contractów Przebudowa presenterScope (dodanie SupervisorJob) Rozbicie pobierania danych z FB na get i snapshot listener Wyodrębnienie sortowania i filtrowania do osobnej funkcji

+++ core/src/main/java/com/rafalskrzypczyk/core/extensions/IdGeneratorIntExt.kt
@@ -2,8 +2,8 @@ package com.rafalskrzypczyk.core.extensions
 
 import java.util.Calendar
 
-fun Int.Companion.generateId() : Int {
+fun Long.Companion.generateId() : Long {
     val timestamp = Calendar.getInstance().timeInMillis
     val randomPart = (0..100).random()
-    return timestamp.toInt() + randomPart
+    return timestamp + randomPart
 }
